refactor: move source into src
